split up transaction logic for readability
